--- web/apps/photos/src/components/Collections/CollectionShare/publicShare/manage/downloadAccess.tsx
@@ -1,57 +1,0 @@
-import type {
-    Collection,
-    PublicURL,
-    UpdatePublicURL,
-} from "@/media/collection";
-import { useAppContext } from "@/new/photos/types/context";
-import { EnteMenuItem } from "@ente/shared/components/Menu/EnteMenuItem";
-import { t } from "i18next";
-import { Trans } from "react-i18next";
-interface Iprops {
-    publicShareProp: PublicURL;
-    collection: Collection;
-    updatePublicShareURLHelper: (req: UpdatePublicURL) => Promise<void>;
-}
-
-export function ManageDownloadAccess({
-    publicShareProp,
-    updatePublicShareURLHelper,
-    collection,
-}: Iprops) {
-    const { showMiniDialog } = useAppContext();
-
-    const handleFileDownloadSetting = () => {
-        if (publicShareProp.enableDownload) {
-            disableFileDownload();
-        } else {
-            updatePublicShareURLHelper({
-                collectionID: collection.id,
-                enableDownload: true,
-            });
-        }
-    };
-
-    const disableFileDownload = () => {
-        showMiniDialog({
-            title: t("disable_file_download"),
-            message: <Trans i18nKey={"disable_file_download_message"} />,
-            continue: {
-                text: t("disable"),
-                color: "critical",
-                action: () =>
-                    updatePublicShareURLHelper({
-                        collectionID: collection.id,
-                        enableDownload: false,
-                    }),
-            },
-        });
-    };
-    return (
-        <EnteMenuItem
-            checked={publicShareProp?.enableDownload ?? true}
-            onClick={handleFileDownloadSetting}
-            variant="toggle"
-            label={t("FILE_DOWNLOAD")}
-        />
-    );
-}
