--- web/apps/photos/src/components/Collections/CollectionShare/publicShare/manage/publicCollect.tsx
@@ -1,42 +1,0 @@
-import { MenuItemGroup, MenuSectionTitle } from "@/base/components/Menu";
-import type {
-    Collection,
-    PublicURL,
-    UpdatePublicURL,
-} from "@/media/collection";
-import { EnteMenuItem } from "@ente/shared/components/Menu/EnteMenuItem";
-import { Stack } from "@mui/material";
-import { t } from "i18next";
-
-interface Iprops {
-    publicShareProp: PublicURL;
-    collection: Collection;
-    updatePublicShareURLHelper: (req: UpdatePublicURL) => Promise<void>;
-}
-
-export function ManagePublicCollect({
-    publicShareProp,
-    updatePublicShareURLHelper,
-    collection,
-}: Iprops) {
-    const handleFileDownloadSetting = () => {
-        updatePublicShareURLHelper({
-            collectionID: collection.id,
-            enableCollect: !publicShareProp.enableCollect,
-        });
-    };
-
-    return (
-        <Stack>
-            <MenuItemGroup>
-                <EnteMenuItem
-                    onClick={handleFileDownloadSetting}
-                    variant="toggle"
-                    checked={publicShareProp?.enableCollect}
-                    label={t("PUBLIC_COLLECT")}
-                />
-            </MenuItemGroup>
-            <MenuSectionTitle title={t("PUBLIC_COLLECT_SUBTEXT")} />
-        </Stack>
-    );
-}
